requirements.txt
# Dependencias para LTM Migration con Dataform
# BigQuery y Google Cloud
google-cloud-bigquery>=6.0.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0

# Dataform CLI
dataform>=1.0.0

# Utilidades de datos
pandas>=1.5.0
numpy>=1.21.0

# Utilidades de desarrollo
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0

# Testing y calidad
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging y monitoreo
structlog>=22.0.0 